move files into separate directory
