+++ packages/server/src/migration_logic/README.md
@@ -4,9 +4,6 @@ npx migrate-mongo create `name`
 
 # apply all pending migrations locally
 
-by starting the server
-
-or by command:
 npx migrate-mongo up
 
 # revert last migration locally
